code made cleaner and still working
