+++ src/main/java/com/greenfox/tribes1/Resources/Food.java
@@ -2,12 +2,10 @@ package com.greenfox.tribes1.Resources;
 
 import lombok.Getter;
 import lombok.Setter;
-import lombok.SneakyThrows;
 import org.springframework.stereotype.Component;
 
 import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
-import java.sql.Timestamp;
 
 @Entity(name = "Food")
 @DiscriminatorValue("Food")
@@ -16,29 +14,9 @@ import java.sql.Timestamp;
 @Component
 public class Food extends Resource {
 
-  //TODO: remove @Transient somehow and make code cleaner
-  //TODO: figure out how to remove TimeService
-
-  public Food(Long amount) {
-    this.setAmount(amount);
-    this.setAmountPerMinute(10L);
-  }
-
   public Food() {
     this.setAmountPerMinute(8L);
   }
-
-  @Override
-  @SneakyThrows
-  public void update(Long difference) {
-    System.out.println(getAmountPerMinute());
-    System.out.println(difference);
-    System.out.println(getAmount());
-    if (difference > 0) {
-      setAmount(getAmountPerMinute() * difference
-              + getAmount());
-      setUpdated_at(new Timestamp(System.currentTimeMillis()));
-    }
-  }
-
 }
+
+
